Crate ic_cdk_macros

This crate provide a set of attribute macros to faciliate canister development.

The macros fall into two categories:

To register functions as canister entry points

To import another canister as a rust struct for inter-canister operation.

Register functions as canister entry points

These macros are directly related to the Internet Computer Specification.

init

pre_upgrade

post_upgrade

inspect_message

heartbeat

update

query

Import another canister as a rust struct

import

Attribute Macros

heartbeat

Register the canister_heartbeat entry point of a canister.

import

Import another canister as a rust struct.

init

Register the canister_init entry point of a canister.

inspect_message

Register the canister_inspect_message entry point of a canister.

post_upgrade

Register the canister_post_upgrade entry point of a canister.

pre_upgrade

Register the canister_pre_upgrade entry point of a canister.

query

Register a query call entry point.

update

Register an update call entry point.